backend/requirements.txt
Django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
boto3==1.34.69
python-decouple==3.8
bcrypt==4.1.3
